requirements.txt
asgiref==3.4.1
attrs==21.2.0
black==21.9b0
click==8.0.1
Django==3.2.6
django-filter==2.4.0
djangorestframework==3.12.4
drf-extensions==0.7.1
factory-boy==3.2.0
Faker==8.12.1
importlib-metadata==4.8.1
inflection==0.5.1
iniconfig==1.1.1
mypy-extensions==0.4.3
packaging==21.0
pathspec==0.9.0
platformdirs==2.4.0
pluggy==1.0.0
psycopg2-binary==2.9.1
py==1.10.0
pyparsing==2.4.7
pytest==6.2.5
pytest-django==4.4.0
pytest-factoryboy==2.1.0
python-dateutil==2.8.2
pytz==2021.1
regex==2021.9.30
six==1.16.0
sqlparse==0.4.1
text-unidecode==1.3
toml==0.10.2
tomli==1.2.1
typed-ast==1.4.3
typing-extensions==3.10.0.0
zipp==3.5.0
